+++ src/Steps/Description/Fields/ConfigurationPlan.tsx
@@ -1,16 +1,5 @@
-import { z } from "zod";
 import BACTextarea from "components/BaseFormFields/BACTextarea";
 import { CAFTOPDescription } from "api/CAFTOP";
-
-const baseRule = z.string().trim();
-
-export const configurationplanRuleSave = z.object({
-  ConfigurationPlan: baseRule,
-});
-
-export const configurationplanRuleFinal = z.object({
-  ConfigurationPlan: baseRule.min(1, "You must enter a Configuration Plan"),
-});
 
 const ConfigurationPlan = () => {
   return (
