+++ src/Steps/Description/Fields/Description.Validation.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+const finalRule = z.string().trim().min(1, "You must enter a Description");
+
+export const DescriptionRuleFinal = z.object({
+  Description: finalRule,
+});
