+++ src/Steps/Description/Fields/Description.tsx
@@ -1,12 +1,5 @@
-import { z } from "zod";
 import BACTextarea from "components/BaseFormFields/BACTextarea";
 import { CAFTOPDescription } from "api/CAFTOP";
-
-const finalRule = z.string().trim().min(1, "You must enter a Description");
-
-export const DescriptionRuleFinal = z.object({
-  Description: finalRule,
-});
 
 const Description = () => {
   return (
