+++ src/Steps/Description/Fields/Introduction.Validation.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+const finalRule = z.string().trim().min(1, "You must enter a Introduction");
+
+export const IntroductionRuleFinal = z.object({
+  Introduction: finalRule,
+});
