+++ src/Steps/Description/Fields/Introduction.tsx
@@ -1,12 +1,5 @@
-import { z } from "zod";
 import BACTextarea from "components/BaseFormFields/BACTextarea";
 import { CAFTOPDescription } from "api/CAFTOP";
-
-const finalRule = z.string().trim().min(1, "You must enter a Introduction");
-
-export const IntroductionRuleFinal = z.object({
-  Introduction: finalRule,
-});
 
 const Introduction = () => {
   return (
