+++ src/Steps/Description/Fields/SystemMissionDescription.tsx
@@ -1,19 +1,5 @@
-import { z } from "zod";
 import BACTextarea from "components/BaseFormFields/BACTextarea";
 import { CAFTOPDescription } from "api/CAFTOP";
-
-const baseRule = z.string().trim();
-
-export const systemmissiondescriptionRuleSave = z.object({
-  SystemMissionDescription: baseRule,
-});
-
-export const systemmissiondescriptionRuleFinal = z.object({
-  SystemMissionDescription: baseRule.min(
-    1,
-    "You must enter a System/Mission Description"
-  ),
-});
 
 const SystemMissionDescription = () => {
   return (
