+++ src/Steps/Distribution/Fields/DSO.tsx
@@ -1,24 +1,6 @@
-import { z } from "zod";
 import { CAFTOPDistribution } from "api/CAFTOP";
 import { Radio } from "@fluentui/react-components";
 import BACRadioGroup from "components/BaseFormFields/BACRadioGroup";
-import { Distribution } from "stateManagement/reducer";
-import { populateWithDefaultValue } from "utilities/Validation";
-
-/** Rule for when the DSO is Not Applicable */
-export const dsoRuleNA = z.object({
-  hasDSO: populateWithDefaultValue(Distribution.hasDSO),
-});
-
-export const dsoRuleSave = z.object({
-  hasDSO: z.enum(["no", "yes", ""]),
-});
-
-export const dsoRuleFinal = z.object({
-  hasDSO: z.enum(["no", "yes"], {
-    message: "You must select whether or not DSO is utiilized",
-  }),
-});
 
 const DSO = () => {
   return (
