+++ src/Steps/Info/Fields/Center.Validation.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+const finalRule = z
+  .string()
+  .trim()
+  .min(1, "You must select a Center from the list");
+
+export const CenterRuleFinal = z.object({
+  Center: finalRule,
+});
