+++ src/Steps/Info/Fields/LeadCommand.tsx
@@ -1,17 +1,7 @@
-import { z } from "zod";
 import BACDropdown from "components/BaseFormFields/BACDropdown";
 import { CAFTOPInfo } from "api/CAFTOP";
 import { Option } from "@fluentui/react-components";
 import { useLeadCommands } from "api/LeadCommands";
-
-const finalRule = z
-  .string()
-  .trim()
-  .min(1, "You must select a Lead Command from the list");
-
-export const LeadCommandRuleFinal = z.object({
-  LeadCommand: finalRule,
-});
 
 export const LeadCommand = () => {
   const LeadCommands = useLeadCommands();
