+++ src/Steps/Info/Fields/PreparingBase.Validation.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+const finalRule = z
+  .string()
+  .trim()
+  .min(1, "You must select a Preparing Base from the list");
+
+export const PreparingBaseRuleFinal = z.object({
+  PreparingBase: finalRule,
+});
