+++ src/Steps/Info/Fields/PreparingBase.tsx
@@ -1,17 +1,7 @@
-import { z } from "zod";
 import BACDropdown from "components/BaseFormFields/BACDropdown";
 import { CAFTOPInfo } from "api/CAFTOP";
 import { Option } from "@fluentui/react-components";
 import { usePreparingBases } from "api/PreparingBases";
-
-const finalRule = z
-  .string()
-  .trim()
-  .min(1, "You must select a Preparing Base from the list");
-
-export const PreparingBaseRuleFinal = z.object({
-  PreparingBase: finalRule,
-});
 
 export const PreparingBase = () => {
   const PreparingBases = usePreparingBases();
