+++ src/Steps/Info/Fields/ProgramElementCode.Validation.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+const finalRule = z
+  .string()
+  .trim()
+  .min(1, "You must select a PEC from the list");
+
+export const ProgramElementCodeRuleFinal = z.object({
+  ProgramElementCode: finalRule,
+});
