+++ src/Steps/Info/Fields/ProgramElementCode.tsx
@@ -1,18 +1,8 @@
-import { z } from "zod";
 import BACDropdown from "components/BaseFormFields/BACDropdown";
 import { CAFTOPInfo } from "api/CAFTOP";
 import { useProgramNamesAndECs } from "api/ProgramNamesAndElementCodes";
 import { Option } from "@fluentui/react-components";
 import { useWatch } from "react-hook-form";
-
-const finalRule = z
-  .string()
-  .trim()
-  .min(1, "You must select a PEC from the list");
-
-export const ProgramElementCodeRuleFinal = z.object({
-  ProgramElementCode: finalRule,
-});
 
 export const ProgramElementCode = () => {
   const ProgramNames = useProgramNamesAndECs();
