+++ src/Steps/Info/Fields/ProgramGroup.tsx
@@ -1,16 +1,6 @@
-import { z } from "zod";
 import BACCombobox from "components/BaseFormFields/BACCombobox";
 import { CAFTOPInfo } from "api/CAFTOP";
 import { useProgramGroups } from "api/ProgramGroups";
-
-const finalRule = z
-  .string()
-  .trim()
-  .min(1, "You must select a Program Group from the list");
-
-export const ProgramGroupRuleFinal = z.object({
-  ProgramGroup: finalRule,
-});
 
 export const ProgramGroup = () => {
   const ProgramGroups = useProgramGroups();
