+++ src/Steps/Info/Fields/ProgramName.Validation.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+const finalRule = z
+  .string()
+  .trim()
+  .min(1, "You must select a Program Name from the list");
+
+export const ProgramNameRuleFinal = z.object({
+  ProgramName: finalRule,
+});
