+++ src/Steps/LRDP/Fields/LRDP.tsx
@@ -1,4 +1,3 @@
-import { z } from "zod";
 import BACInput from "components/BaseFormFields/BACInput";
 import { CAFTOPLRDP } from "api/CAFTOP";
 import { useFieldArray, useWatch } from "react-hook-form";
@@ -16,7 +15,6 @@ import {
   Text,
 } from "@fluentui/react-components";
 import BACRadioGroup from "components/BaseFormFields/BACRadioGroup";
-import { LRDP as LRDPDefault } from "stateManagement/reducer";
 import {
   DeleteRegular,
   ArrowUp16Regular,
@@ -24,88 +22,6 @@ import {
   DragRegular,
 } from "@fluentui/react-icons";
 import { SyntheticEvent, DragEvent } from "react";
-import { populateWithDefaultValue } from "utilities/Validation";
-
-const lrdpName = z.string().trim().max(50, "Name cannot exceed 50 characters");
-const lrdpSeqNumFinal = z
-  .string()
-  .trim()
-  .regex(/^\d{5}$/, "This must be a 5 digits");
-const lrdpSeqNumSave = z.union([z.literal(""), lrdpSeqNumFinal]);
-
-const checkForDuplicates = (items: CAFTOPLRDP, ctx: z.RefinementCtx) => {
-  const uniqueValuesName = new Map<string, number>();
-  const uniqueValuesSeqNum = new Map<string, number>();
-  items.LRDP.forEach((item, idx) => {
-    const lowerCaseName = item.Name.toLowerCase().trim(); // Make case insensitive and remove spacing
-    const firstAppearanceNameIndex = uniqueValuesName.get(lowerCaseName);
-    const firstAppearanceSeqNumIndex = uniqueValuesSeqNum.get(item.SeqNum);
-    if (firstAppearanceNameIndex !== undefined && lowerCaseName !== "") {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: `Name must be unique`,
-        path: ["LRDP", idx, "Name"],
-      });
-    } else {
-      uniqueValuesName.set(lowerCaseName, idx);
-    }
-    if (firstAppearanceSeqNumIndex !== undefined && item.SeqNum !== "") {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: `Sequence Number must be unique`,
-        path: ["LRDP", idx, "SeqNum"],
-      });
-    } else {
-      uniqueValuesSeqNum.set(item.SeqNum, idx);
-    }
-  });
-};
-
-export const lrdpRuleSave = z
-  .discriminatedUnion("hasLRDP", [
-    z.object({
-      hasLRDP: z.enum(["no", ""]),
-      LRDP: populateWithDefaultValue(LRDPDefault.LRDP),
-    }),
-    z.object({
-      hasLRDP: z.literal("yes"),
-      LRDP: z.array(z.object({ Name: lrdpName, SeqNum: lrdpSeqNumSave })),
-    }),
-  ])
-  .superRefine(checkForDuplicates);
-
-export const lrdpRuleFinal = z
-  .discriminatedUnion(
-    "hasLRDP",
-    [
-      z.object({
-        hasLRDP: z.enum(["no"]),
-        LRDP: populateWithDefaultValue(LRDPDefault.LRDP),
-      }),
-      z.object({
-        hasLRDP: z.literal("yes"),
-        LRDP: z
-          .array(
-            z.object({
-              Name: lrdpName.min(1, "You must supply a name"),
-              SeqNum: lrdpSeqNumFinal,
-            })
-          )
-          .min(1, "You must have at least one LRDP priority"),
-      }),
-    ],
-    {
-      errorMap: (issue, ctx) => {
-        if (issue.code === z.ZodIssueCode.invalid_union_discriminator) {
-          return {
-            message: "You must select whether there are LRDP tasks or not",
-          };
-        }
-        return { message: ctx.defaultError };
-      },
-    }
-  )
-  .superRefine(checkForDuplicates);
 
 function allowDrop(e: SyntheticEvent) {
   e.preventDefault();
