+++ src/Steps/Labor/Fields/LaborType.Validation.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+const finalRule = z.string().trim().min(1, "You must select a Labor Type");
+
+export const LaborTypeRuleFinal = z.object({
+  LaborType: finalRule,
+});
