+++ src/Steps/Labor/Fields/LaborType.tsx
@@ -1,13 +1,6 @@
-import { z } from "zod";
 import BACRadioGroup from "components/BaseFormFields/BACRadioGroup";
 import { CAFTOPLabor } from "api/CAFTOP";
 import { Radio } from "@fluentui/react-components";
-
-const finalRule = z.string().trim().min(1, "You must select a Labor Type");
-
-export const LaborTypeRuleFinal = z.object({
-  LaborType: finalRule,
-});
 
 const LaborType = () => {
   return (
